Treat p-values as numeric before sorting
The p variable at some point became a list of character strings that when sorted, do not sort properly due to the number of digits in the p-value. For instance, a p-value of 0.999984 would be sorted as being larger than 0.9999864 even though it's numeric value is smaller. 

I believe this has the effect of increasing the q-values from the qvalue function because the list is no longer in order, so larger q-values may sometimes be treated as being smaller. This will then decrease the number of significant associations output by WGCNA.